refactor(explorer): use typed NodeKind/EdgeKind in C4 inference (E6.5) (#14)
* refactor(explorer): use typed NodeKind/EdgeKind in C4 inference (E6.5)

Sprint E6.5 type-safety. Closes the partial status from
docs/explorer-roadmap.md: E6.5.

The build_architecture/compare_architecture methods used plain string
literals for kind and relation fields. Replaced with typed enum calls:

  "system"     → NodeKind::System.as_str().to_string()
  "container"  → NodeKind::Container.as_str().to_string()
  "component"  → NodeKind::Component.as_str().to_string()
  "part_of"    → EdgeKind::PartOf.as_str().to_string()

The "code" string is kept as a literal — it's a C4-specific mapping
(SymbolKind → "code" when in a C4 component) that is NOT part of
the global NodeKind enum. Verified by the
build_architecture_emits_code_kind_for_c4_code_nodes test.

Implementation refactored:
- Extracted build_architecture_impl behind #[cfg(feature = "multimodal")]
- Without multimodal: returns ExplorerError::FeatureDisabled
- With multimodal: uses typed enum values

Tests gated behind #[cfg(all(test, feature = "multimodal"))] since
they require the C4 inference code.

Wire format unchanged (still strings).
No breaking changes.

Tag: v0.12.11 PATCH

* fix(explorer): resolve pre-existing multimodal build errors

These errors blocked the E6.5 type-safety refactor from being validated
under --features multimodal. Each is a 1-3 line fix:

1. api.rs: SubgraphResponse + GraphRepository not in scope
   - Added to top-level use statements (line 13, line 22)
   - Removed redundant function-scoped import

2. api_rationale_tests.rs: WorkspaceService + ViewService not in scope
   - Added to use crate::facades::{...} imports
   - Fixed GraphServiceImpl import path (was crate::facades::, now crate::facades::graph::)

3. mcp/handler/ingest.rs: MAX_GRAPH_SEARCH_LIMIT type mismatch
   - Cast to usize to match DEFAULT_GRAPH_SEARCH_LIMIT arm

4. mcp/handler/sessions.rs: explicit ref binding modifier
   - Removed redundant ref in if let (Some(name), Some(k)) = (...)

Verified: cargo check -p cognicode-explorer --features multimodal exits 0
with the E6.5 refactor + these fixes. Tag: v0.12.11 PATCH

---------

+++ crates/cognicode-explorer/src/mcp/handler/ingest.rs
@@ -207,8 +207,8 @@ impl ToolHandler for GraphSearchHandler {
                 return err_envelope(TOOL_GRAPH_SEARCH, "invalid_input",
                     &format!("graph_search: `limit` must be a positive integer (got {n})"));
             }
-            Some(n) => (n as usize).min(MAX_GRAPH_SEARCH_LIMIT),
-            None => DEFAULT_GRAPH_SEARCH_LIMIT,
+            Some(n) => (n as usize).min(MAX_GRAPH_SEARCH_LIMIT as usize),
+            None => DEFAULT_GRAPH_SEARCH_LIMIT as usize,
         };
 
         let mut parsed_kinds: Vec<NodeKind> = Vec::new();
